refactor: move main function
